+++ operator-framework-core/src/main/java/io/javaoperatorsdk/operator/processing/event/NamedEventSource.java
@@ -4,7 +4,7 @@ import java.util.Objects;
 import java.util.Optional;
 
 import io.javaoperatorsdk.operator.OperatorException;
-import io.javaoperatorsdk.operator.api.reconciler.Reconciler;
+import io.javaoperatorsdk.operator.api.reconciler.EventSourceUtils;
 import io.javaoperatorsdk.operator.processing.event.source.Configurable;
 import io.javaoperatorsdk.operator.processing.event.source.EventSource;
 import io.javaoperatorsdk.operator.processing.event.source.EventSourceStartPriority;
@@ -19,7 +19,7 @@ class NamedEventSource implements EventSource, EventSourceMetadata {
   NamedEventSource(EventSource original, String name) {
     this.original = original;
     this.name = name;
-    nameSet = !name.equals(Reconciler.generateNameFor(original));
+    nameSet = !name.equals(EventSourceUtils.generateNameFor(original));
   }
 
   @Override
