+++ operator-framework-core/src/test/java/io/javaoperatorsdk/operator/api/reconciler/EventSourceInitializerTest.java
@@ -15,8 +15,8 @@ class EventSourceInitializerTest {
   void defaultNameDifferentForOtherInstance() {
     var eventSource1 = new PollingEventSource(HashMap::new, 1000, String.class);
     var eventSource2 = new PollingEventSource(HashMap::new, 1000, String.class);
-    var eventSourceName1 = Reconciler.generateNameFor(eventSource1);
-    var eventSourceName2 = Reconciler.generateNameFor(eventSource2);
+    var eventSourceName1 = EventSourceUtils.generateNameFor(eventSource1);
+    var eventSourceName2 = EventSourceUtils.generateNameFor(eventSource2);
 
     assertThat(eventSourceName1).isNotEqualTo(eventSourceName2);
   }
