+++ operator-framework/src/test/java/io/javaoperatorsdk/operator/sample/indexdiscriminator/IndexDiscriminatorTestReconciler.java
@@ -81,7 +81,7 @@ public class IndexDiscriminatorTestReconciler
     secondDependentResourceConfigMap
         .setResourceDiscriminator(
             new TestIndexDiscriminator(CONFIG_MAP_INDEX_2, FIRST_CONFIG_MAP_SUFFIX_2));
-    return Reconciler.nameEventSources(eventSource);
+    return EventSourceUtils.nameEventSources(eventSource);
   }
 
   public static String configMapKey(ConfigMap configMap) {
