+++ operator-framework/src/test/java/io/javaoperatorsdk/operator/sample/multipledependentresource/MultipleDependentResourceReconciler.java
@@ -64,6 +64,6 @@ public class MultipleDependentResourceReconciler
     firstDependentResourceConfigMap.configureWith(eventSource);
     secondDependentResourceConfigMap.configureWith(eventSource);
 
-    return Reconciler.nameEventSources(eventSource);
+    return EventSourceUtils.nameEventSources(eventSource);
   }
 }
