+++ operator-framework/src/test/java/io/javaoperatorsdk/operator/sample/primarytosecondary/JobReconciler.java
@@ -79,7 +79,7 @@ public class JobReconciler
               primary.getSpec().getClusterName(), primary.getMetadata().getNamespace())));
     }
 
-    return Reconciler
+    return EventSourceUtils
         .nameEventSources(new InformerEventSource<>(informerConfiguration.build(), context));
   }
